chore: eslint config: rename folder and remove unnecessary import
